+++ src/App.jsx
@@ -5,9 +5,10 @@ import Header from "./components/Header/Header";
 import RecipeList from "./components/RecipeList/RecipeList";
 import Recipe from "./components/recipe/Recipe";
 import { useRecipes } from "./hooks/useRecipes";
-import AddRecipePage from "./pages/AddRecipe/AddRecipePage";
+import AddRecipePage from "./pages/AddRecipe/AddRecipe";
 import CategoryFilter from "./components/CategoryFilter/CategoryFilter";
 import { Squirrel } from "lucide-react";
+import EditRecipePage from "./pages/AddRecipe/EditRecipe";
 
 const defaultCategories = [
   "Alle Rezepte",
@@ -56,6 +57,10 @@ function App() {
         <Route
           path="/add-recipe"
           element={<AddRecipePage categories={defaultCategories} />}
+        />
+        <Route
+          path="/edit-recipe"
+          element={<EditRecipePage categories={defaultCategories} />}
         />
       </Routes>
     </Router>
