+++ src/components/Recipe/Recipe.jsx
@@ -1,10 +1,12 @@
-import { useParams } from "react-router-dom";
+import { useParams, useNavigate } from "react-router-dom";
 import { useRecipe } from "../../hooks/useRecipe";
 import "./Recipe.css";
+import { Pencil, ShoppingBasket } from "lucide-react";
 
 const Recipe = () => {
   const { slug } = useParams();
   const { recipe, loading, error } = useRecipe(slug);
+  const navigate = useNavigate();
 
   if (loading) return <div>Loading...</div>; // TODO - add loading animation after a certain amount of time
   if (error) return <div>Error: {error}</div>;
@@ -12,7 +14,17 @@ const Recipe = () => {
 
   return (
     <div className="recipe-container">
-      <h1 className="recipe-title">{recipe.title}</h1>
+      <div className="recipe-heading">
+        <h1 className="recipe-title">{recipe.title}</h1>
+        <button
+          className="edit-btn"
+          onClick={() => {
+            navigate("/edit-recipe");
+          }}
+        >
+          <Pencil />
+        </button>
+      </div>
       {recipe.servings && (
         <div style={{ display: "flex", alignItems: "center", gap: "0.5em" }}>
           <h3>Servings:</h3>
@@ -22,17 +34,28 @@ const Recipe = () => {
       {recipe.ingredients && recipe.ingredients.length > 0 && (
         <>
           <h3>Ingredients:</h3>
-          <ul>
+          <ul style={{ listStyle: "none", padding: 0 }}>
             {recipe.ingredients.map((ingredient) => (
-              <li key={ingredient.id}>
-                {ingredient.quantity && `${ingredient.quantity} `}
-                {ingredient.unit && `${ingredient.unit} `}
-                {ingredient.name}
+              <li
+                key={ingredient.id}
+                style={{ display: "flex", alignItems: "center", gap: "0.5em" }}
+              >
+                <input type="checkbox" id={`ingredient-${ingredient.id}`} />
+                <label htmlFor={`ingredient-${ingredient.id}`}>
+                  {ingredient.quantity && `${ingredient.quantity} `}
+                  {ingredient.unit && `${ingredient.unit} `}
+                  {ingredient.name}
+                </label>
               </li>
             ))}
           </ul>
         </>
       )}
+
+      <button className="cart-btn">
+        {/* TODO - add a counter in the top right of the icon to keep trcak of how many items in cart */}
+        <ShoppingBasket size={28} />
+      </button>
 
       {recipe.instructions && (
         <>
