+++ src/pages/AddRecipe/EditRecipe.jsx
@@ -1,0 +1,5 @@
+const EditRecipePage = ({ categories }) => {
+  // TODO
+};
+
+export default EditRecipePage;
